+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Use the context API in TYPO3 9.5 for the FE login (#363, #366)
 - Streamline the login managers (#365)
 
 ## 3.0.2